fix bug

+++ controllers/books.js
@@ -1,30 +1,69 @@
 const Book = require('../models/book');
+const fs = require('fs');
 
 exports.createBook = (req, res, next) => {
+  const bookObject = JSON.parse(req.body.book);
+  delete bookObject._id;
+  delete bookObject._userId;
+  delete bookObject.ratings._id;
+  delete bookObject.averageRating._id;
   const book = new Book({
-     ...req.body
+      ...bookObject,
+      userId: req.auth.userId,
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
   });
+
+  console.log(JSON.stringify(book));
   book.save()
-     .then(() =>  res.status(201).json({ message : "Livre créé avec succès" }))
-     .catch(error => res.status(400).json({ error }));
-}
+  .then(() => { res.status(201).json({message: 'Le livre a été enregistré !'})})
+  .catch(error => { res.status(400).json( { error: error.toString() })})
+};
 
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id : req.params.id})
-      .then(()=>res.status(200).json({message:"Le livre a été supprimé"}))
-      .catch(err=>res.status(400).json({error: err}))
-}
+  Book.findOne({ _id: req.params.id})
+      .then(book => {
+          if (book.userId != req.auth.userId) {
+              res.status(401).json({message: 'Not authorized'});
+          } else {
+              const filename = book.imageUrl.split('/images/')[1];
+              fs.unlink(`images/${filename}`, () => {
+                  Book.deleteOne({_id: req.params.id})
+                      .then(() => { res.status(200).json({message: 'Le livre a été supprimé !'})})
+                      .catch(error => res.status(401).json({ error }));
+              });
+          }
+      })
+      .catch( error => {
+          res.status(500).json({ error });
+      });
+};
 
 exports.updateBook = (req, res, next) => {
-  Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Livre modifié !'}))
-    .catch(error => res.status(400).json({ error }));
-}
+  const bookObject = req.file ? {
+      ...JSON.parse(req.body.book),
+      imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+  } : { ...req.body };
+
+  delete bookObject._userId;
+  Book.findOne({_id: req.params.id})
+      .then((book) => {
+          if (book.userId != req.auth.userId) {
+              res.status(401).json({ message : 'Not authorized'});
+          } else {
+              Book.updateOne({ _id: req.params.id}, { ...bookObject, _id: req.params.id})
+              .then(() => res.status(200).json({message : 'Le livre a été modifié!'}))
+              .catch(error => res.status(401).json({ error }));
+          }
+      })
+      .catch((error) => {
+          res.status(400).json({ error });
+      });
+};
 
 exports.getOneBook = (req, res, next) => {
   Book.findOne({ _id: req.params.id })
     .then(book => res.status(200).json(book))
-    .catch(error => res.status(404).json({ error }));
+    .catch(error => res.status(404).json({ error: error.toString() }));
 }
 
 exports.getAllBooks = (req, res, next) => {
